BLD: add legal terms + intro page
