+++ pages/thank-you.js
@@ -68,7 +68,7 @@ function ThankYou(props) {
   }, []);
 
   return (
-    <Container component="main" maxWidth="sm">
+    <Container component="main" maxWidth="sm" className={classes.root}>
       <Head>
         <title>Thank you</title>
       </Head>
