+++ styles/global.css
@@ -44,3 +44,9 @@ footer a {
 footer a:hover {
 	text-decoration: underline;
 }
+
+.disclaimer {
+	font-size: 0.9em;
+	color: darkgrey;
+	font-style: italic;
+}
